+++ akka-docs-new/src/test/java/jdocs/tutorial_4/DeviceTest.java
@@ -57,7 +57,7 @@ public class DeviceTest extends JUnitSuite {
     ActorRef deviceActor = system.actorOf(Device.props("group", "device"));
     deviceActor.tell(new Device.ReadTemperature(42L), probe.getRef());
     Device.RespondTemperature response = probe.expectMsgClass(Device.RespondTemperature.class);
-    Assert.assertEquals((Long) 42L, response.requestId);
+    Assert.assertEquals(42L, response.requestId);
     Assert.assertEquals(Optional.empty(), response.value);
   }
 
@@ -67,19 +67,19 @@ public class DeviceTest extends JUnitSuite {
     ActorRef deviceActor = system.actorOf(Device.props("group", "device"));
 
     deviceActor.tell(new Device.RecordTemperature(1L, 24.0), probe.getRef());
-    Assert.assertEquals((Long) 1L, probe.expectMsgClass(Device.TemperatureRecorded.class).requestId);
+    Assert.assertEquals(1L, probe.expectMsgClass(Device.TemperatureRecorded.class).requestId);
 
     deviceActor.tell(new Device.ReadTemperature(2L), probe.getRef());
     Device.RespondTemperature response1 = probe.expectMsgClass(Device.RespondTemperature.class);
-    Assert.assertEquals((Long) 2L, response1.requestId);
+    Assert.assertEquals(2L, response1.requestId);
     Assert.assertEquals(Optional.of(24.0), response1.value);
 
     deviceActor.tell(new Device.RecordTemperature(3L, 55.0), probe.getRef());
-    Assert.assertEquals((Long) 3L, probe.expectMsgClass(Device.TemperatureRecorded.class).requestId);
+    Assert.assertEquals(3L, probe.expectMsgClass(Device.TemperatureRecorded.class).requestId);
 
     deviceActor.tell(new Device.ReadTemperature(4L), probe.getRef());
     Device.RespondTemperature response2 = probe.expectMsgClass(Device.RespondTemperature.class);
-    Assert.assertEquals((Long) 4L, response2.requestId);
+    Assert.assertEquals(4L, response2.requestId);
     Assert.assertEquals(Optional.of(55.0), response2.value);
   }
