+++ akka-docs-new/src/test/java/jdocs/tutorial_5/DeviceGroupTest.java
@@ -52,9 +52,9 @@ public class DeviceGroupTest extends JUnitSuite {
 
     // Check that the device actors are working
     deviceActor1.tell(new Device.RecordTemperature(0L, 1.0), probe.getRef());
-    Assert.assertEquals((Long) 0L, probe.expectMsgClass(Device.TemperatureRecorded.class).requestId);
+    Assert.assertEquals(0L, probe.expectMsgClass(Device.TemperatureRecorded.class).requestId);
     deviceActor2.tell(new Device.RecordTemperature(1L, 2.0), probe.getRef());
-    Assert.assertEquals((Long) 1L, probe.expectMsgClass(Device.TemperatureRecorded.class).requestId);
+    Assert.assertEquals(1L, probe.expectMsgClass(Device.TemperatureRecorded.class).requestId);
   }
 
   @Test
@@ -94,7 +94,7 @@ public class DeviceGroupTest extends JUnitSuite {
 
     groupActor.tell(new DeviceGroup.RequestDeviceList(0L), probe.getRef());
     DeviceGroup.ReplyDeviceList reply = probe.expectMsgClass(DeviceGroup.ReplyDeviceList.class);
-    Assert.assertEquals((Long) 0L, reply.requestId);
+    Assert.assertEquals(0L, reply.requestId);
     Assert.assertEquals(Stream.of("device1", "device2").collect(Collectors.toSet()), reply.ids);
   }
 
@@ -112,7 +112,7 @@ public class DeviceGroupTest extends JUnitSuite {
 
     groupActor.tell(new DeviceGroup.RequestDeviceList(0L), probe.getRef());
     DeviceGroup.ReplyDeviceList reply = probe.expectMsgClass(DeviceGroup.ReplyDeviceList.class);
-    Assert.assertEquals((Long) 0L, reply.requestId);
+    Assert.assertEquals(0L, reply.requestId);
     Assert.assertEquals(Stream.of("device1", "device2").collect(Collectors.toSet()), reply.ids);
 
     probe.watch(toShutDown);
@@ -121,7 +121,7 @@ public class DeviceGroupTest extends JUnitSuite {
 
     groupActor.tell(new DeviceGroup.RequestDeviceList(1L), probe.getRef());
     reply = probe.expectMsgClass(DeviceGroup.ReplyDeviceList.class);
-    Assert.assertEquals((Long) 1L, reply.requestId);
+    Assert.assertEquals(1L, reply.requestId);
     Assert.assertEquals(Stream.of("device2").collect(Collectors.toSet()), reply.ids);
   }
 
@@ -144,14 +144,14 @@ public class DeviceGroupTest extends JUnitSuite {
 
     // Check that the device actors are working
     deviceActor1.tell(new Device.RecordTemperature(0L, 1.0), probe.getRef());
-    Assert.assertEquals((Long) 0L, probe.expectMsgClass(Device.TemperatureRecorded.class).requestId);
+    Assert.assertEquals(0L, probe.expectMsgClass(Device.TemperatureRecorded.class).requestId);
     deviceActor2.tell(new Device.RecordTemperature(1L, 2.0), probe.getRef());
-    Assert.assertEquals((Long) 1L, probe.expectMsgClass(Device.TemperatureRecorded.class).requestId);
+    Assert.assertEquals(1L, probe.expectMsgClass(Device.TemperatureRecorded.class).requestId);
     // No temperature for device 3
 
     groupActor.tell(new DeviceGroup.RequestAllTemperatures(0L), probe.getRef());
     DeviceGroup.RespondAllTemperatures response = probe.expectMsgClass(DeviceGroup.RespondAllTemperatures.class);
-    Assert.assertEquals((Long) 0L, response.requestId);
+    Assert.assertEquals(0L, response.requestId);
 
     Map<String, DeviceGroup.TemperatureReading> expectedTemperatures = new HashMap<>();
     expectedTemperatures.put("device1", new DeviceGroup.Temperature(1.0));
